Parallelize GTTA API for improving performance
DO NOT MERGE

With the numbers of GTTA requests incoming, the response time is
getting longer and longer since the GTTA function is executed one
at a time.
By removing the 'synchronized' keyword, the GTTA becomes parallel
and the performance is improved. The throughtput is increased and
the response time becomes shorter.

Experiment result reference: https://github.com/iotaledger/iri/issues/973.